fix: modify Piece for upgrading
We need to keep track of the number of rounds each Piece survives. When
they survive a specified number of rounds, we upgrade their rarity and
stats. We also need to keep track of their initial rarity, so that when
we sell the piece back, it goes back to their correct rarity pool.